fix: router type define
