+++ packages/runtime/plugin-router-legacy/.eslintrc.js
@@ -1,4 +1,8 @@
 module.exports = {
   root: true,
   extends: ['@modern-js'],
+  parserOptions: {
+    tsconfigRootDir: __dirname,
+    project: ['./tsconfig.json'],
+  },
 };
