+++ packages/runtime/plugin-router-legacy/tsconfig.json
@@ -3,7 +3,7 @@
   "compilerOptions": {
     "declaration": false,
     "jsx": "preserve",
-    "baseUrl": ".",
+    "baseUrl": "./",
     "isolatedModules": true,
     "paths": {}
   },
