--- packages/runtime/plugin-runtime/type.d.ts
@@ -1,5 +1,0 @@
-declare module '@modern-js/runtime' {
-  export * from './dist/types';
-}
-
-export * from './dist/types';
