+++ packages/runtime/plugin-runtime/types/index.d.ts
@@ -1,0 +1,13 @@
+import { StateConfig } from '../dist/types';
+
+declare module '@modern-js/core/config' {
+  interface RuntimeConfig {
+    state?: StateConfig | boolean;
+  }
+}
+
+declare module 'http' {
+  interface ServerResponse {
+    locals: Record<string, any>;
+  }
+}
